requirements.txt
pillow
piexif
pikepdf
python-docx
openpyxl
python-pptx
ffmpeg-python
hachoir
